dont use float

+++ src/main.c
@@ -394,8 +394,8 @@ run:
                 timeout = 0;
             }
 
-            seltv.tv_sec = (long)((float)timeout / 1E3f);
-            seltv.tv_nsec = (long)(1E6f*((float)timeout - 1E3f*(float)seltv.tv_sec));
+            seltv.tv_sec = (long)(timeout / 1E3f);
+            seltv.tv_nsec = (long)(1E6f*(timeout - 1E3f*(float)seltv.tv_sec));
             if (timeout >= 0) {
                 tv = &seltv;
             } else {
@@ -443,7 +443,7 @@ run:
             if (CONF_BLINK_TIMEOUT && term_attr_set(ATTR_BLINK)) {
                 timeout = CONF_BLINK_TIMEOUT - timediff(now, lastblink);
                 if (timeout <= 0) {
-                    if (-timeout > (float)CONF_BLINK_TIMEOUT) {
+                    if (-timeout > CONF_BLINK_TIMEOUT) {
                         term_window.mode |= WIN_MODE_BLINK;
                     }
                     term_window.mode ^= WIN_MODE_BLINK;
